Look for screen option the right way first and fall back to request variable or a reasonable default
If there's a screen option set, we want to use that. But if we're
loading this data with ajax, the page is `admin-ajax.php` and doesn't
have access to options. So let's check for a screen option, if that
doesn't work let them pass in a number with `$_REQUEST['per_page']` and
finally fall back to a reasonable default of 20.